Add some of the Botan test cases for QBigInteger - lots more cases still to be done.
My machine is delivering failing tests for a few steps (currently 7
failures) and 17 passes. There are a lot of tests commented out -
any use of a QBigInteger( QString ) constructor with the QString
representing a negative number causes SIGABRT.

Of the 7 cases that are failing, two are for subtraction. Looks like
-= is broken.

svn path=/trunk/kdesupport/qca/; revision=357243