Ensure that we can match on subject altnames, as well as the CN.
Also update the documentation to indicate the behaviour.

There might be a case for adding a bool argument to matchesHostname()
so that you can test if it matches just the CN, or either CN or
altname. 

This function should probably also test IP address (and perhaps also
XMPP) altnames, but I'd like to think about that a bit more first, and
trial it with a real certificate.

CCMAIL: justin@affinix.com
CCBUG: 107604


svn path=/trunk/kdesupport/qca/; revision=540564